Gateway/requirements.txt
fastapi
fastapi[standard]
uvicorn[standard]
httpx
python-jose[cryptography]
passlib[bcrypt]
pydantic-settings # For loading .env
slowapi
redis[hiredis]
# email-validator # Often needed by FastAPI for EmailStr
# python-multipart # For handling form data like in login